Convert PDF to HTML

This is wrapper around API to cloudconvert.

convert_pdf_2_html(
  input_file_url,
  output_file = NULL,
  quiet = FALSE,
  api_key = NULL,
  overwrite = FALSE
)

Arguments

input_file_url
A URL to pdf file to process. Do not supply path to file in your filesystem.

output_file
A name of file to write .html to. Must be of .html extension. By default, the name of the file is extracted from input_file_url.

quiet
Boolean. Should the method display messages about progress? Errors will be displayed regardless of this option.

api_key
A custom API key obtained from cloudconvert. Make sure to use the key of version 1.

overwrite
Boolean. Should the .html file be overwritten?

Value

A HTML file with name output_file is created. It is ready to be passed into extract_code_from_html function.

Details

Create your own account at cloudconvert and obtain your own API key, or convert the file(s) manually.

See also